requirements.txt
python-dotenv>=0.19.2
flask
requests
Flask-Session